+++ python/lsst/ts/cRIOpy/GUI/SAL/TopicData.py
@@ -1,18 +1,43 @@
-from lsst.ts.cRIOpy.M1M3FATable import (
-    FATABLE_XINDEX,
-    FATABLE_YINDEX,
-    FATABLE_ZINDEX,
-    FATABLE_SINDEX,
-)
+# This file is part of the cRIO GUI.
+#
+# Developed for the LSST Telescope and Site Systems.
+# This product includes software developed by the LSST Project
+# (https://www.lsst.org). See the COPYRIGHT file at the top - level directory
+# of this distribution for details of code ownership.
+#
+# This program is free software : you can redistribute it and / or modify it
+# under the terms of the GNU General Public License as published by the Free
+# Software Foundation, either version 3 of the License, or (at your option) any
+# later version.
+#
+# This program is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+# FOR A PARTICULAR PURPOSE.See the GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License along with
+# this program.If not, see <https://www.gnu.org/licenses/>.
 
 from ..ActuatorsDisplay import Scales
 
-__all__ = ["Topics", "TopicData", "TopicField"]
+__all__ = ["TopicData", "TopicField"]
 
 
 class TopicData:
     """
-    Helper class. Used together with TopicField to construct topics.
+    Holds topic available for display.
+
+    Attributes
+    ----------
+    name : `str`
+        Name to display in selection list box.
+    fields : `[TopicField]`
+        Fields available inside topic. Array of TopicField.
+    topic : `str`
+        Name of the topics. Equals to SAL/DDS topic name for telemetry, needs
+        evt_ prefix for events.
+    isEvent : `bool`
+        True if topic is an event. Data are extracted from remote with evt_
+        prefix.
     """
 
     def __init__(self, name, fields, topic, isEvent=True):
@@ -29,835 +54,41 @@ class TopicData:
 
 
 class TopicField:
-    def __init__(self, name, value, valueIndex, scale=Scales.GAUGE):
+    """
+    Field inside topic. Holds together data needed to properly display values.
+
+    Attributes
+    ----------
+    name : `str`
+        Text displayed in selection list box.
+    fieldName : `str`
+        Name of field inside SAL/DDS topic.
+    valueIndex : `int`
+        Variable kind.
+    scaleType : `Scales`
+        Scale type. Select scale used to display field values.
+    """
+
+    def __init__(self, name, fieldName, valueIndex, scaleType=Scales.GAUGE):
         self.name = name
-        self.value = value
+        self.fieldName = fieldName
         self.valueIndex = valueIndex
-        self.scale = scale
+        self.scaleType = scaleType
 
     def getValue(self, data):
-        return getattr(data, self.value)
+        return getattr(data, self.fieldName)
 
 
 class OnOffField(TopicField):
-    def __init__(self, name, value, valueIndex):
-        super().__init__(name, value, valueIndex, Scales.ONOFF)
+    def __init__(self, name, fieldName, valueIndex):
+        super().__init__(name, fieldName, valueIndex, Scales.ONOFF)
 
 
 class WarningField(TopicField):
-    def __init__(self, name, value, valueIndex):
-        super().__init__(name, value, valueIndex, Scales.WARNING)
-
-
-class BumpTestField(TopicField):
-    def __init__(self, name, value, valueIndex):
-        super().__init__(name, value, valueIndex, Scales.BUMP_TEST)
+    def __init__(self, name, fieldName, valueIndex):
+        super().__init__(name, fieldName, valueIndex, Scales.WARNING)
 
 
 class EnabledDisabledField(TopicField):
-    def __init__(self, name, value, valueIndex):
-        super().__init__(name, value, valueIndex, Scales.ENABLED_DISABLED)
-
-
-class Topics:
-    """
-    Class constructing list of all available topics.
-    """
-
-    def __init__(self):
-        self.lastIndex = None
-
-        self.topics = [
-            TopicData(
-                "Applied Acceleration Forces",
-                [
-                    TopicField("Z Forces", "zForces", FATABLE_ZINDEX),
-                    TopicField("Y Forces", "yForces", FATABLE_YINDEX),
-                    TopicField("X Forces", "xForces", FATABLE_XINDEX),
-                ],
-                "appliedAccelerationForces",
-                False,
-            ),
-            TopicData(
-                "Applied Active Optic Forces",
-                [TopicField("Z Forces", "zForces", FATABLE_ZINDEX)],
-                "appliedActiveOpticForces",
-            ),
-            TopicData(
-                "Applied Azimuth Forces",
-                [
-                    TopicField("Z Forces", "zForces", FATABLE_ZINDEX),
-                    TopicField("Y Forces", "yForces", FATABLE_YINDEX),
-                    TopicField("X Forces", "xForces", FATABLE_XINDEX),
-                ],
-                "appliedAzimuthForces",
-                False,
-            ),
-            TopicData(
-                "Applied Balance Forces",
-                [
-                    TopicField("Z Forces", "zForces", FATABLE_ZINDEX),
-                    TopicField("Y Forces", "yForces", FATABLE_YINDEX),
-                    TopicField("X Forces", "xForces", FATABLE_XINDEX),
-                ],
-                "appliedBalanceForces",
-                False,
-            ),
-            TopicData(
-                "Applied Cylinder Forces",
-                [
-                    TopicField(
-                        "Primary Cylinder Forces",
-                        "primaryCylinderForces",
-                        FATABLE_ZINDEX,
-                    ),
-                    TopicField(
-                        "Secondary Cylinder Forces",
-                        "secondaryCylinderForces",
-                        FATABLE_SINDEX,
-                    ),
-                ],
-                "appliedCylinderForces",
-                False,
-            ),
-            TopicData(
-                "Applied Elevation Forces",
-                [
-                    TopicField("Z Forces", "zForces", FATABLE_ZINDEX),
-                    TopicField("Y Forces", "yForces", FATABLE_YINDEX),
-                    TopicField("X Forces", "xForces", FATABLE_XINDEX),
-                ],
-                "appliedElevationForces",
-                False,
-            ),
-            TopicData(
-                "Applied Forces",
-                [
-                    TopicField("Z Forces", "zForces", FATABLE_ZINDEX),
-                    TopicField("Y Forces", "yForces", FATABLE_YINDEX),
-                    TopicField("X Forces", "xForces", FATABLE_XINDEX),
-                ],
-                "appliedForces",
-                False,
-            ),
-            TopicData(
-                "Applied Offset Forces",
-                [
-                    TopicField("Z Forces", "zForces", FATABLE_ZINDEX),
-                    TopicField("Y Forces", "yForces", FATABLE_YINDEX),
-                    TopicField("X Forces", "xForces", FATABLE_XINDEX),
-                ],
-                "appliedOffsetForces",
-            ),
-            TopicData(
-                "Applied Static Forces",
-                [
-                    TopicField("Z Forces", "zForces", FATABLE_ZINDEX),
-                    TopicField("Y Forces", "yForces", FATABLE_YINDEX),
-                    TopicField("X Forces", "xForces", FATABLE_XINDEX),
-                ],
-                "appliedStaticForces",
-            ),
-            TopicData(
-                "Applied Thermal Forces",
-                [
-                    TopicField("Z Forces", "zForces", FATABLE_ZINDEX),
-                    TopicField("Y Forces", "yForces", FATABLE_YINDEX),
-                    TopicField("X Forces", "xForces", FATABLE_XINDEX),
-                ],
-                "appliedThermalForces",
-                False,
-            ),
-            TopicData(
-                "Applied Velocity Forces",
-                [
-                    TopicField("Z Forces", "zForces", FATABLE_ZINDEX),
-                    TopicField("Y Forces", "yForces", FATABLE_YINDEX),
-                    TopicField("X Forces", "xForces", FATABLE_XINDEX),
-                ],
-                "appliedVelocityForces",
-                False,
-            ),
-            TopicData(
-                "Pre-clipped Acceleration Forces",
-                [
-                    TopicField("Z Forces", "zForces", FATABLE_ZINDEX),
-                    TopicField("Y Forces", "yForces", FATABLE_YINDEX),
-                    TopicField("X Forces", "xForces", FATABLE_XINDEX),
-                ],
-                "preclippedAccelerationForces",
-            ),
-            TopicData(
-                "Pre-clipped Active Optic Forces",
-                [TopicField("Z Forces", "zForces", FATABLE_ZINDEX)],
-                "preclippedActiveOpticForces",
-            ),
-            TopicData(
-                "Pre-clipped Azimuth Forces",
-                [
-                    TopicField("Z Forces", "zForces", FATABLE_ZINDEX),
-                    TopicField("Y Forces", "yForces", FATABLE_YINDEX),
-                    TopicField("X Forces", "xForces", FATABLE_XINDEX),
-                ],
-                "preclippedAzimuthForces",
-            ),
-            TopicData(
-                "Pre-clipped Balance Forces",
-                [
-                    TopicField("Z Forces", "zForces", FATABLE_ZINDEX),
-                    TopicField("Y Forces", "yForces", FATABLE_YINDEX),
-                    TopicField("X Forces", "xForces", FATABLE_XINDEX),
-                ],
-                "preclippedBalanceForces",
-            ),
-            TopicData(
-                "Pre-clipped Cylinder Forces",
-                [
-                    TopicField(
-                        "Primary Cylinder Forces",
-                        "primaryCylinderForces",
-                        FATABLE_ZINDEX,
-                    ),
-                    TopicField(
-                        "Secondary Cylinder Forces",
-                        "secondaryCylinderForces",
-                        FATABLE_SINDEX,
-                    ),
-                ],
-                "preclippedCylinderForces",
-            ),
-            TopicData(
-                "Pre-clipped Elevation Forces",
-                [
-                    TopicField("Z Forces", "zForces", FATABLE_ZINDEX),
-                    TopicField("Y Forces", "yForces", FATABLE_YINDEX),
-                    TopicField("X Forces", "xForces", FATABLE_XINDEX),
-                ],
-                "preclippedElevationForces",
-            ),
-            TopicData(
-                "Pre-clipped Forces",
-                [
-                    TopicField("Z Forces", "zForces", FATABLE_ZINDEX),
-                    TopicField("Y Forces", "yForces", FATABLE_YINDEX),
-                    TopicField("X Forces", "xForces", FATABLE_XINDEX),
-                ],
-                "preclippedForces",
-            ),
-            TopicData(
-                "Pre-clipped Offset Forces",
-                [
-                    TopicField("Z Forces", "zForces", FATABLE_ZINDEX),
-                    TopicField("Y Forces", "yForces", FATABLE_YINDEX),
-                    TopicField("X Forces", "xForces", FATABLE_XINDEX),
-                ],
-                "preclippedOffsetForces",
-            ),
-            TopicData(
-                "Pre-clipped Static Forces",
-                [
-                    TopicField("Z Forces", "zForces", FATABLE_ZINDEX),
-                    TopicField("Y Forces", "yForces", FATABLE_YINDEX),
-                    TopicField("X Forces", "xForces", FATABLE_XINDEX),
-                ],
-                "preclippedStaticForces",
-            ),
-            TopicData(
-                "Pre-clipped Thermal Forces",
-                [
-                    TopicField("Z Forces", "zForces", FATABLE_ZINDEX),
-                    TopicField("Y Forces", "yForces", FATABLE_YINDEX),
-                    TopicField("X Forces", "xForces", FATABLE_XINDEX),
-                ],
-                "preclippedThermalForces",
-            ),
-            TopicData(
-                "Pre-clipped Velocity Forces",
-                [
-                    TopicField("Z Forces", "zForces", FATABLE_ZINDEX),
-                    TopicField("Y Forces", "yForces", FATABLE_YINDEX),
-                    TopicField("X Forces", "xForces", FATABLE_XINDEX),
-                ],
-                "preclippedVelocityForces",
-            ),
-            TopicData(
-                "Measured forces",
-                [
-                    TopicField(
-                        "Primary Cylinder Forces",
-                        "primaryCylinderForce",
-                        FATABLE_ZINDEX,
-                    ),
-                    TopicField(
-                        "Secondary Cylinder Forces",
-                        "secondaryCylinderForce",
-                        FATABLE_SINDEX,
-                    ),
-                    TopicField(
-                        "Z Forces",
-                        "zForce",
-                        FATABLE_ZINDEX,
-                    ),
-                    TopicField(
-                        "Y Forces",
-                        "yForce",
-                        FATABLE_YINDEX,
-                    ),
-                    TopicField(
-                        "X Forces",
-                        "xForce",
-                        FATABLE_XINDEX,
-                    ),
-                ],
-                "forceActuatorData",
-                False,
-            ),
-            TopicData(
-                "FA Following Error",
-                [
-                    TopicField(
-                        "Primary Cylinder FE",
-                        "primaryCylinderFollowingError",
-                        FATABLE_ZINDEX,
-                    ),
-                    TopicField(
-                        "Secondary Cylinder FE",
-                        "secondaryCylinderFollowingError",
-                        FATABLE_SINDEX,
-                    ),
-                ],
-                "forceActuatorData",
-                False,
-            ),
-            TopicData(
-                "FA ILC Info",
-                [
-                    TopicField("Subnet", "modbusSubnet", FATABLE_ZINDEX),
-                    TopicField("Address", "modbusAddress", FATABLE_ZINDEX),
-                    TopicField(
-                        "Major Revision",
-                        "majorRevision",
-                        FATABLE_ZINDEX,
-                    ),
-                    TopicField(
-                        "Minor Revision",
-                        "minorRevision",
-                        FATABLE_ZINDEX,
-                    ),
-                    TopicField("ADC Scan Rate", "adcScanRate", FATABLE_ZINDEX),
-                ],
-                "forceActuatorInfo",
-            ),
-            TopicData(
-                "FA Id Info",
-                [
-                    TopicField(
-                        "X Data Reference Id",
-                        "xDataReferenceId",
-                        FATABLE_XINDEX,
-                    ),
-                    TopicField(
-                        "Y Data Reference Id",
-                        "yDataReferenceId",
-                        FATABLE_YINDEX,
-                    ),
-                    TopicField(
-                        "Z Data Reference Id",
-                        "zDataReferenceId",
-                        FATABLE_ZINDEX,
-                    ),
-                    TopicField(
-                        "S Data Reference Id",
-                        "sDataReferenceId",
-                        FATABLE_SINDEX,
-                    ),
-                    TopicField("ILC Unique Id", "ilcUniqueId", FATABLE_ZINDEX),
-                    TopicField(
-                        "Mezzanine Unique Id",
-                        "xDataReferenceId",
-                        FATABLE_ZINDEX,
-                    ),
-                ],
-                "forceActuatorInfo",
-            ),
-            TopicData(
-                "FA Settings",
-                [
-                    EnabledDisabledField(
-                        "Enabled/Disabled", "enabledActuators", FATABLE_ZINDEX
-                    ),
-                    TopicField(
-                        "PC FE Warning",
-                        "primaryFollowingErrorWarningThreshold",
-                        FATABLE_ZINDEX,
-                    ),
-                    TopicField(
-                        "PC FE Counting",
-                        "primaryFollowingErrorCountingFaultThreshold",
-                        FATABLE_ZINDEX,
-                    ),
-                    TopicField(
-                        "PC FE Immediate",
-                        "primaryFollowingErrorImmediateFaultThreshold",
-                        FATABLE_ZINDEX,
-                    ),
-                    TopicField(
-                        "SC FE Warning",
-                        "secondaryFollowingErrorWarningThreshold",
-                        FATABLE_SINDEX,
-                    ),
-                    TopicField(
-                        "SC FE Counting",
-                        "secondaryFollowingErrorCountingFaultThreshold",
-                        FATABLE_SINDEX,
-                    ),
-                    TopicField(
-                        "SC FE Immediate",
-                        "secondaryFollowingErrorImmediateFaultThreshold",
-                        FATABLE_SINDEX,
-                    ),
-                ],
-                "forceActuatorSettings",
-            ),
-            TopicData(
-                "FA Main Calibration Info",
-                [
-                    TopicField(
-                        "Primary Coefficient",
-                        "mainPrimaryCylinderCoefficient",
-                        FATABLE_ZINDEX,
-                    ),
-                    TopicField(
-                        "Primary Offset",
-                        "mainPrimaryCylinderLoadCellOffset",
-                        FATABLE_ZINDEX,
-                    ),
-                    TopicField(
-                        "Primary Sensitivity",
-                        "mainPrimaryCylinderLoadCellSensitivity",
-                        FATABLE_ZINDEX,
-                    ),
-                    TopicField(
-                        "Secondary Coefficient",
-                        "mainSecondaryCylinderCoefficient",
-                        FATABLE_SINDEX,
-                    ),
-                    TopicField(
-                        "Secondary Offset",
-                        "mainSecondaryCylinderLoadCellOffset",
-                        FATABLE_SINDEX,
-                    ),
-                    TopicField(
-                        "Secondary Sensitivity",
-                        "mainSecondaryLoadCellSensitivity",
-                        FATABLE_SINDEX,
-                    ),
-                ],
-                "forceActuatorInfo",
-            ),
-            TopicData(
-                "FA Backup Calibration Info",
-                [
-                    TopicField(
-                        "Primary Coefficient",
-                        "backupPrimaryCylinderCoefficient",
-                        FATABLE_ZINDEX,
-                    ),
-                    TopicField(
-                        "Primary Offset",
-                        "backupPrimaryCylinderLoadCellOffset",
-                        FATABLE_ZINDEX,
-                    ),
-                    TopicField(
-                        "Primary Sensitivity",
-                        "backupPrimaryCylinderLoadCellSensitivity",
-                        FATABLE_ZINDEX,
-                    ),
-                    TopicField(
-                        "Secondary Coefficient",
-                        "backupSecondaryCylinderCoefficient",
-                        FATABLE_SINDEX,
-                    ),
-                    TopicField(
-                        "Secondary Offset",
-                        "backupSecondaryCylinderLoadCellOffset",
-                        FATABLE_SINDEX,
-                    ),
-                    TopicField(
-                        "Secondary Sensitivity",
-                        "backupSecondaryLoadCellSensitivity",
-                        FATABLE_SINDEX,
-                    ),
-                ],
-                "forceActuatorInfo",
-            ),
-            TopicData(
-                "FA Mezzanine Calibration Info",
-                [
-                    TopicField(
-                        "Primary Cylinder Gain",
-                        "mezzaninePrimaryCylinderGain",
-                        FATABLE_ZINDEX,
-                    ),
-                    TopicField(
-                        "Secondary Cylinder Gain",
-                        "mezzanineSecondaryCylinderGain",
-                        FATABLE_SINDEX,
-                    ),
-                ],
-                "forceActuatorInfo",
-            ),
-            TopicData(
-                "FA Position Info",
-                [
-                    TopicField(
-                        "Actuator Type",
-                        "actuatorType",
-                        FATABLE_ZINDEX,
-                    ),
-                    TopicField(
-                        "Actuator Orientation",
-                        "actuatorOrientation",
-                        FATABLE_ZINDEX,
-                    ),
-                    TopicField("X Position", "xPosition", FATABLE_ZINDEX),
-                    TopicField("Y Position", "yPosition", FATABLE_ZINDEX),
-                    TopicField("Z Position", "zPosition", FATABLE_ZINDEX),
-                ],
-                "forceActuatorInfo",
-            ),
-            TopicData(
-                "FA State",
-                [TopicField("ILC State", "ilcState", FATABLE_ZINDEX)],
-                "forceActuatorState",
-            ),
-            TopicData(
-                "FA Warning",
-                [
-                    WarningField("Major Fault", "majorFault", FATABLE_ZINDEX),
-                    WarningField("Minor Fault", "minorFault", FATABLE_ZINDEX),
-                    WarningField(
-                        "Fault Override",
-                        "faultOverride",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Main Calibration Error",
-                        "mainCalibrationError",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Backup Calibration Error",
-                        "backupCalibrationError",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Mezzanine Error",
-                        "mezzanineError",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Mezzanine Bootloader Active",
-                        "mezzanineBootloaderActive",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Unique Id CRC Error",
-                        "uniqueIdCRCError",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Application Type Mismatch",
-                        "applicationTypeMismatch",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Application Missing",
-                        "applicationMissing",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "OneWire Mismatch",
-                        "oneWireMissing",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "OneWire1 Mismatch",
-                        "oneWire1Mismatch",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "OnewWire2 Mismatch",
-                        "oneWire2Mismatch",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Watchdog Reset",
-                        "watchdogReset",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Brownout",
-                        "brownOut",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Event Trap Reset",
-                        "eventTrapReset",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "SSR Power Fault",
-                        "ssrPowerFault",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "AUX Power Fault",
-                        "auxPowerFault",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Mezzanine Power Fault",
-                        "mezzaninePowerFault",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Mezzanine Current Ampl Fault",
-                        "mezzanineCurrentAmp1Fault",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Mezzanine Current Amp2 Fault",
-                        "mezzanineCurrentAmp2Fault",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Mezzanine Unique ID CRC Error",
-                        "mezzanineUniqueIdCRCError",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Mezzanine Main Calibration Error",
-                        "mezzanineMainCalibrationError",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Mezzanine Backup Calibration Error",
-                        "mezzanineBackupCalibrationError",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Mezzanine Event Trap Reset",
-                        "mezzanineEventTrapReset",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Mezzanine Application Missing",
-                        "mezzanineApplicationMissing",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Mezzanine Application CRC Mismatch",
-                        "mezzanineApplicationCRCMismatch",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "ILC Fault",
-                        "ilcFault",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Broadcast Counter Warning",
-                        "broadcastCounterWarning",
-                        FATABLE_ZINDEX,
-                    ),
-                ],
-                "forceActuatorWarning",
-            ),
-            TopicData(
-                "FA Force Warning",
-                [
-                    WarningField(
-                        "Primary Axis Measured Force Warning",
-                        "primaryAxisMeasuredForceWarning",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Secondary Axis Measured Force Warning",
-                        "secondaryAxisMeasuredForceWarning",
-                        FATABLE_SINDEX,
-                    ),
-                    WarningField(
-                        "Primary Axis FE Warning",
-                        "primaryAxisFollowingErrorWarning",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Secondary Axis FE Warning",
-                        "secondaryAxisFollowingErrorWarning",
-                        FATABLE_SINDEX,
-                    ),
-                    WarningField(
-                        "Primary Axis FE Counting Fault",
-                        "primaryAxisFollowingErrorCountingFault",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Secondary Axis FE Counting Fault",
-                        "secondaryAxisFollowingErrorCountingFault",
-                        FATABLE_SINDEX,
-                    ),
-                    WarningField(
-                        "Primary Axis FE Immediate Fault",
-                        "primaryAxisFollowingErrorImmediateFault",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Secondary Axis FE Immediate Fault",
-                        "secondaryAxisFollowingErrorImmediateFault",
-                        FATABLE_SINDEX,
-                    ),
-                ],
-                "forceActuatorForceWarning",
-            ),
-            TopicData(
-                "FA FE Counters",
-                [
-                    TopicField(
-                        "Primary axis FE warning counter",
-                        "primaryAxisFollowingErrorWarningCounter",
-                        FATABLE_ZINDEX,
-                    ),
-                    TopicField(
-                        "Secondary axis FE warning counter",
-                        "secondaryAxisFollowingErrorWarningCounter",
-                        FATABLE_SINDEX,
-                    ),
-                    TopicField(
-                        "Primary axis FE counting counter",
-                        "primaryAxisFollowingErrorCountingCounter",
-                        FATABLE_ZINDEX,
-                    ),
-                    TopicField(
-                        "Secondary axis FE counting counter",
-                        "secondaryAxisFollowingErrorCountingCounter",
-                        FATABLE_SINDEX,
-                    ),
-                ],
-                "forceActuatorFollowingErrorCounter",
-            ),
-            TopicData(
-                "FA Setpoint Warning",
-                [
-                    WarningField(
-                        "Safety limit",
-                        "safetyLimitWarning",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Near neighbor",
-                        "nearNeighborWarning",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Far neighbor",
-                        "farNeighborWarning",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Elevation force",
-                        "elevationForceWarning",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Azimuth force",
-                        "azimuthForceWarning",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Thermal force",
-                        "thermalForceWarning",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Balance force",
-                        "balanceForceWarning",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Acceleration force",
-                        "accelerationForceWarning",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Active optic",
-                        "activeOpticForceWarning",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Static force",
-                        "staticForceWarning",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Offset force",
-                        "offsetForceWarning",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Velocity force",
-                        "velocityForceWarning",
-                        FATABLE_ZINDEX,
-                    ),
-                    WarningField(
-                        "Force setpoint",
-                        "forceWarning",
-                        FATABLE_ZINDEX,
-                    ),
-                ],
-                "forceSetpointWarning",
-            ),
-            TopicData(
-                "FA Bump Test",
-                [
-                    BumpTestField("Primary Test", "primaryTest", FATABLE_ZINDEX),
-                    BumpTestField(
-                        "Secondary Test",
-                        "secondaryTest",
-                        FATABLE_SINDEX,
-                    ),
-                    TopicField(
-                        "Primary Timestamps",
-                        "primaryTestTimestamps",
-                        FATABLE_ZINDEX,
-                    ),
-                    TopicField(
-                        "Secondary Timestamps",
-                        "secondaryTestTimestamps",
-                        FATABLE_SINDEX,
-                    ),
-                ],
-                "forceActuatorBumpTestStatus",
-            ),
-            TopicData(
-                "FA enabled",
-                [
-                    EnabledDisabledField(
-                        "Enabled FAs",
-                        "forceActuatorEnabled",
-                        FATABLE_ZINDEX,
-                    ),
-                ],
-                "enabledForceActuators",
-            ),
-        ]
-
-    def changeTopic(self, index, slot, comm):
-        if self.lastIndex is not None:
-            getattr(comm, self.topics[self.lastIndex].topic).disconnect(slot)
-
-        self.lastIndex = index
-        if index is None:
-            return
-
-        getattr(comm, self.topics[index].topic).connect(slot)
+    def __init__(self, name, fieldName, valueIndex):
+        super().__init__(name, fieldName, valueIndex, Scales.ENABLED_DISABLED)
